requirements.txt
google-generativeai==0.5.4
streamlit==1.35.0